refactor: 패키지 구조 정리 및 Registration 도메인 분리
