+++ src/main/java/nextstep/courses/domain/base/BaseInfo.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.base;
 
 import java.time.LocalDateTime;
