+++ src/main/java/nextstep/courses/domain/batch/BatchInfo.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.batch;
 
 public class BatchInfo {
