+++ src/main/java/nextstep/courses/domain/course/CourseInfo.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.course;
 
 public class CourseInfo {
