+++ src/main/java/nextstep/courses/domain/curriculum/Curriculums.java
@@ -1,8 +1,9 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.curriculum;
 
 import java.util.HashSet;
 import java.util.Set;
 import nextstep.courses.DuplicatedException;
+import nextstep.courses.domain.curriculum.Curriculum;
 
 public class Curriculums {
