+++ src/main/java/nextstep/courses/domain/registration/RegistrationRepository.java
@@ -1,6 +1,7 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.registration;
 
 import java.util.Optional;
+import nextstep.courses.domain.registration.Registration;
 
 public interface RegistrationRepository {
