+++ src/main/java/nextstep/courses/domain/registration/Registrations.java
@@ -1,7 +1,8 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.registration;
 
 import java.util.HashSet;
 import java.util.Set;
+import nextstep.courses.domain.registration.Registration;
 import nextstep.users.domain.NsUser;
 
 public class Registrations {
