+++ src/main/java/nextstep/courses/domain/session/SessionRepository.java
@@ -1,6 +1,7 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import java.util.Optional;
+import nextstep.courses.domain.session.Session;
 
 public interface SessionRepository {
