+++ src/test/java/nextstep/courses/domain/BatchTest.java
@@ -4,6 +4,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDateTime;
 import nextstep.courses.DuplicatedException;
+import nextstep.courses.domain.batch.Batch;
+import nextstep.courses.domain.course.Course;
+import nextstep.courses.domain.session.Session;
+import nextstep.courses.domain.session.SessionType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
