+++ src/test/java/nextstep/courses/domain/RegistrationTest.java
@@ -8,6 +8,10 @@ import java.time.LocalDateTime;
 import nextstep.courses.DuplicatedException;
 import nextstep.courses.RegistrationFulledException;
 import nextstep.courses.RegistrationNotOpenedException;
+import nextstep.courses.domain.batch.Batch;
+import nextstep.courses.domain.registration.Registration;
+import nextstep.courses.domain.session.Session;
+import nextstep.courses.domain.session.SessionType;
 import nextstep.users.domain.NsUser;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
